Restructure the software for debugging hardware
The decoder now provides a sequence of decoded blocks.

The aim is for the hardware testbenches to pull out software decoded
blocks and check against the hardware results showing any
differences as they occur.

To make this work, we will still need to modify the software dct to
match the hardware.